Add tests (integration/unit test), go pkg changes
