+++ .changelog/Unreleased/10.md
@@ -1,0 +1,5 @@
+# Changed
+
+- Adjusted newlines around generated changelog.
+- `chalog` package refactored to expose only a `Generator` interface and a `NewGenerator` function. The `Generator`
+struct implementation is no longer exposed.
