+++ go.mod
@@ -4,6 +4,7 @@ go 1.16
 
 require (
 	github.com/Masterminds/semver v1.5.0
+	github.com/google/go-cmp v0.5.5
 	github.com/mattn/go-runewidth v0.0.12
 	github.com/rivo/uniseg v0.2.0 // indirect
 	github.com/yuin/goldmark v1.3.4
