+++ integration/.test_changelog/releases.txt
@@ -1,0 +1,6 @@
+dev
+test
+v1.0.0,2020-12-25
+v0.4.0-alpha,2019-12-25
+v0.4.0,2018-12-25
+v0.0.1,2017-12-25
