+++ integration/.test_changelog/v0.4.0-alpha/1.md
@@ -1,0 +1,4 @@
+# Added
+
+- Test value
+- Test 2
