+++ integration/.test_changelog/v0.4.0-alpha/2.md
@@ -1,0 +1,7 @@
+# Added
+
+- Test
+
+```go
+This is code
+```
