+++ integration/.test_config.yml
@@ -1,0 +1,2 @@
+repo: https://github.com/jthomperoo/test_repo
+in: .test_changelog/
