+++ integration/EXPECTED.MD
@@ -1,0 +1,36 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [dev]
+### Custom
+- This is a custom header
+
+## [test]
+
+
+## [v1.0.0] - 2020-12-25
+
+
+## [v0.4.0-alpha] - 2019-12-25
+### Added
+- Test value
+- Test 2
+- Test
+```go
+This is code
+```
+
+## [v0.4.0] - 2018-12-25
+
+
+## [v0.0.1] - 2017-12-25
+
+
+[dev]: https://github.com/jthomperoo/test_repo/compare/test...dev
+[test]: https://github.com/jthomperoo/test_repo/compare/v1.0.0...test
+[v1.0.0]: https://github.com/jthomperoo/test_repo/compare/v0.4.0-alpha...v1.0.0
+[v0.4.0-alpha]: https://github.com/jthomperoo/test_repo/compare/v0.4.0...v0.4.0-alpha
+[v0.4.0]: https://github.com/jthomperoo/test_repo/compare/v0.0.1...v0.4.0
+[v0.0.1]: https://github.com/jthomperoo/test_repo/releases/tag/v0.0.1
